Fix api module update
