+++ plugin/src/main/kotlin/com/jetbrains/packagesearch/plugin/tests/PKGSCommandProvider.kt
@@ -3,7 +3,7 @@ package com.jetbrains.packagesearch.plugin.tests
 import com.jetbrains.packagesearch.plugin.tests.dumps.DumpPackageSearchModules
 import com.jetbrains.performancePlugin.CommandProvider
 
-class PKGSCommandProvider : CommandProvider {
+internal class PKGSCommandProvider : CommandProvider {
 
     override fun getCommands() = mapOf(
         DumpPackageSearchModules.COMMAND
