+++ plugin/src/test/kotlin/com/jetbrains/packagesearch/plugin/tests/projectservice/MavenProjectsTest.kt
@@ -1,0 +1,10 @@
+package com.jetbrains.packagesearch.plugin.tests.projectservice
+
+import org.junit.jupiter.api.TestInstance
+
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class MavenProjectsTest : PackageSearchProjectServiceTest() {
+
+    override val resourcePath = "/projects/maven/"
+
+}
